# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for income and expense applications that allows you to receive a report for a certain period",
        "title": "Expense Application",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080/api",
    "basePath": "/",
    "paths": {
...
        "/budgets": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Method that return list of budgets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budgets"
                ],
                "operationId": "get-budgets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/BudgetsResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Method that store budget",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budgets"
                ],
                "operationId": "store-budgets",
                "parameters": [
                    {
                        "description": "Budget form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/BudgetCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Status"
                        }
                    }
                }
            }
        },
        "/budgets/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Method that return budget by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budgets"
                ],
                "operationId": "get-budget",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Budget ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/BudgetResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Method that allow to update budget by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budgets"
                ],
                "operationId": "update-budgets",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Budget ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Budget form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/BudgetUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Status"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Method that allow to delete budget by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budgets"
                ],
                "operationId": "delete-budgets",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Budget ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Status"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "BudgetCreateRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 10.25
                },
                "category_slug": {
                    "type": "string",
                    "example": "test-category"
                },
                "title": {
                    "type": "string",
                    "example": "New car"
                },
                "type": {
                    "type": "string",
                    "example": "income"
                },
                "user_id": {
                    "type": "integer",
                    "example": 12
                }
            }
        },
        "BudgetResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 1000
                },
                "id": {
                    "type": "integer",
                    "example": 10
                },
                "title": {
                    "type": "string",
                    "example": "New budget"
                },
                "type": {
                    "type": "string",
                    "example": "income"
                },
                "user_id": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "BudgetUpdateRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 10.25
                },
                "title": {
                    "type": "string",
                    "example": "New car"
                },
                "type": {
                    "type": "string",
                    "example": "income"
                },
                "user_id": {
                    "type": "integer",
                    "example": 12
                }
            }
        },
        "BudgetsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/BudgetResponse"
                    }
                }
            }
        },
...
        "Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Something went wrong try again later!"
                }
            }
        },
...
        "Status": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "Everything is OK"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}